Refactor due to circular imports
